[editor+core] Add field::set<T> + made field::get<T> const.
This is needed to implement field wrappers, which are special kind of
fields that, when set, call some callback.
I'm going to use these to implement an actor instance system, on actor
can be instanced and customized at runtime. I needed some sort of
callback to know which properties of an actor have been modified in an
instance, so i came up with this change.